+++ plugin/app/Services/ReviewCommentService.php
@@ -5,7 +5,9 @@ namespace TTU\Charon\Services;
 use TTU\Charon\Dto\FileReviewCommentsDTO;
 use TTU\Charon\Dto\ReviewCommentDTO;
 use TTU\Charon\Exceptions\ReviewCommentException;
+use TTU\Charon\Models\Charon;
 use TTU\Charon\Repositories\ReviewCommentRepository;
+use TTU\Charon\Repositories\SubmissionsRepository;
 use Zeizig\Moodle\Globals\User;
 
 /**
@@ -18,31 +20,59 @@ class ReviewCommentService
     /** @var ReviewCommentRepository */
     private $reviewCommentRepository;
 
+    /** @var NotificationService */
+    private $notificationService;
+
+    /** @var SubmissionsRepository */
+    private $submissionRepository;
+
     /**
      * ReviewCommentService constructor.
      *
      * @param ReviewCommentRepository $reviewCommentRepository
+     * @param NotificationService $notificationService
+     * @param SubmissionsRepository $submissionRespository
      */
-    public function __construct(ReviewCommentRepository $reviewCommentRepository)
-    {
+    public function __construct(
+        ReviewCommentRepository $reviewCommentRepository,
+        NotificationService $notificationService,
+        SubmissionsRepository $submissionRespository
+    ) {
         $this->reviewCommentRepository = $reviewCommentRepository;
+        $this->notificationService = $notificationService;
+        $this->submissionRepository = $submissionRespository;
     }
 
     /**
      * Get logged-in user's identifier and save their review comment.
      *
-     * @param $submissionFileId
-     * @param $reviewComment
-     * @param $notify
+     * @param int $submissionFileId
+     * @param string $reviewComment
+     * @param bool $notify
+     * @param Charon $charon
      * @throws ReviewCommentException
      */
-    public function add($submissionFileId, $reviewComment, $notify): void
-    {
+    public function add(
+        int $submissionFileId,
+        string $reviewComment,
+        bool $notify,
+        Charon $charon
+    ) {
         if (strlen($reviewComment) > 10000) {
             throw new ReviewCommentException("review_comment_over_limit");
         }
         $userId = app(User::class)->currentUserId();
-        $this->reviewCommentRepository->add($userId, $submissionFileId, $reviewComment, $notify);
+        $comment = $this->reviewCommentRepository->add($userId, $submissionFileId, $reviewComment, $notify);
+        if ($comment && $notify) {
+            $submissionFile = $this->submissionRepository->getSubmissionFileById($submissionFileId);
+            if ($submissionFile) {
+                $this->notificationService->sendNotificationToStudent(
+                    $submissionFile->submission_id,
+                    $reviewComment,
+                    $charon,
+                    $submissionFile->path);
+            }
+        }
     }
 
     /**